Nicco cleaning

+++ MobileApp/MobileApp/Views/ProfilePage.xaml.cs
@@ -1,13 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Text;
 using System.Threading.Tasks;
-using Xamarin.Essentials;
 using Xamarin.Forms;
-using Xamarin.Forms.Internals;
 using Xamarin.Forms.Xaml;
 
 namespace MobileApp.Views
@@ -15,7 +10,6 @@ namespace MobileApp.Views
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class ProfilePage : TabbedPage
     {
-        private string DynamicPrimaryTextColor;
 
         public string UserName { get; set; }
         public string UserEmail { get; set; }
@@ -27,7 +21,7 @@ namespace MobileApp.Views
         {
             InitializeComponent();
             UserName = App.UserName;
-            UserEmail  = App.UserEmail;
+            UserEmail = App.UserEmail;
 
             if (App.CurrentPlayer != null)
             {
@@ -58,14 +52,14 @@ namespace MobileApp.Views
                     new RowDefinition {Height = 10}
                 }
             };
-       
-         
+
+
 
             StackLayout stackLayout = new StackLayout
             {
                 Margin = new Thickness(20, 0, 20, 0),
                 Padding = new Thickness(0, 70, 0, 0)
-                
+
             };
 
             Label labelCampaign = new Label
@@ -75,8 +69,8 @@ namespace MobileApp.Views
                 FontSize = Device.GetNamedSize(NamedSize.Medium, typeof(Label)),
                 Margin = new Thickness(10, 10, 0, 10)
             };
-            
-     
+
+
             Label labelExperience = new Label
             {
                 Text = $"Experience Level: {App.CurrentDM.ExperienceLevel}",
@@ -176,7 +170,7 @@ namespace MobileApp.Views
                 FontSize = Device.GetNamedSize(NamedSize.Medium, typeof(Label)),
                 Margin = new Thickness(10, 10, 0, 10)
             };
-          
+
             Label labelClass = new Label
             {
                 Text = $"Class: {App.CurrentPlayer.Class}",
@@ -239,7 +233,7 @@ namespace MobileApp.Views
             stackLayout.Children.Add(button);
 
             Grid.SetRow(stackLayout, 1);
-           
+
             ProfileGrid.Children.Add(stackLayout);
         }
 
@@ -250,61 +244,58 @@ namespace MobileApp.Views
         /// <returns>Task of completion</returns>
         async Task OnDeleteClicked(Button button)
         {
-            Busy(button);
-            HttpClient client = new HttpClient();
+            var confirmation = await DisplayAlert("", "Are you sure you want to delete your character? There is no way to get them back.", "Just Do It!", "Nope, nevermind.");
+            if (confirmation)
+            {
 
-            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", App.UserToken);
-            bool result = false;
-            if (App.CurrentPlayer != null)
-            {
-                result = await DeletePlayer(client);
-            }
-            else
-            {
-                result = await DeleteDM(client);
-            }
+                Busy(button);
+                HttpClient client = new HttpClient();
 
-            NotBusy(button);
-            if (result)
-            {
-                App.CurrentPlayer = null;
-                App.CurrentDM = null;
-                // True for New Profile
-                // False for logout
-                var choice = await DisplayAlert("Success", "Profile successfully deleted. Would you like to Logout or Create a new Profile?", "New Profile", "Logout");
-                if (choice)
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", App.UserToken);
+                bool result = false;
+                if (App.CurrentPlayer != null)
                 {
-                    // True for DM
-                    // False for Player
-                    bool profileType = await DisplayAlert("", "Would you like to sign up with a Dungeon Master or a Player profile?", "Dungeon Master", "Player");
-                    if (profileType)
+                    result = await DeletePlayer(client);
+                }
+                else
+                {
+                    result = await DeleteDM(client);
+                }
+
+                NotBusy(button);
+                if (result)
+                {
+                    App.CurrentPlayer = null;
+                    App.CurrentDM = null;
+                    // True for New Profile
+                    // False for logout
+                    var choice = await DisplayAlert("Success", "Profile successfully deleted. Would you like to Logout or Create a new Profile?", "New Profile", "Logout");
+                    if (choice)
                     {
-                        Application.Current.MainPage = new NavigationPage(new DMCreationPage());
-
-
-                        // Application.Current.MainPage = new DMCreationPage();
+                        // True for DM
+                        // False for Player
+                        bool profileType = await DisplayAlert("", "Would you like to sign up with a Dungeon Master or a Player profile?", "Dungeon Master", "Player");
+                        if (profileType)
+                        {
+                            await Shell.Current.GoToAsync($"DMCreationPage");
+                        }
+                        else
+                        {
+                            await Shell.Current.GoToAsync($"PlayerCreationPage");
+                        }
                     }
                     else
                     {
-                        Application.Current.MainPage = new NavigationPage(new PlayerCreationPage());
-
-                        //await Shell.Current.GoToAsync("//PlayerCreationPage");
-
-
-                        // Application.Current.MainPage = new PlayerCreationPage();
+                        App.UserToken = null;
+                        App.UserName = null;
+                        App.UserId = null;
+                        Application.Current.MainPage = new NavigationPage(new LoginPage());
                     }
                 }
                 else
                 {
-                    App.UserToken = null;
-                    App.UserName = null;
-                    App.UserId = null;
-                    Application.Current.MainPage = new NavigationPage(new LoginPage());
+                    await DisplayAlert("Oh No!", "Something went wrong deleting your Profile. Please try again.", "OK");
                 }
-            }
-            else
-            {
-                await DisplayAlert("Oh No!", "Something went wrong deleting your Profile. Please try again.", "OK");
             }
         }
 
@@ -316,7 +307,7 @@ namespace MobileApp.Views
         /// <returns>Task of completion with boolean representing if the delete was successful</returns>
         public async Task<bool> DeletePlayer(HttpClient client)
         {
-            HttpResponseMessage result = await client.DeleteAsync($"{App.ApiUrl}/Players");
+            HttpResponseMessage result = await client.DeleteAsync($"{App.ApiUrl}/Players/{App.CurrentPlayer.Id}");
 
             return result.IsSuccessStatusCode;
         }
@@ -328,10 +319,11 @@ namespace MobileApp.Views
         /// <returns>Task of completion with boolean representing if the delete was successful</returns>
         public async Task<bool> DeleteDM(HttpClient client)
         {
-            HttpResponseMessage result = await client.DeleteAsync($"{App.ApiUrl}/DungeonMasters");
+            HttpResponseMessage result = await client.DeleteAsync($"{App.ApiUrl}/DungeonMasters/{App.CurrentDM.Id}");
 
             return result.IsSuccessStatusCode;
         }
+
 
         /// <summary>
         /// Shows activity indicator and disables buttons
